refactor: layout 폴더링 수정
