+++ app/(workspace)/@header/default.tsx
@@ -1,4 +1,4 @@
-import { Header } from '@/widget/layout';
+import { Header } from '@/widget/layout/workspace';
 
 export default function page() {
   return <Header />;
