+++ app/(workspace)/@sidebar/default.tsx
@@ -1,4 +1,4 @@
-import { Sidebar } from '@/widget/layout';
+import { Sidebar } from '@/widget/layout/workspace';
 
 export default function page() {
   return <Sidebar />;
