+++ src/widget/layout/workspace/index.ts
